requirements.txt
google-cloud-aiplatform
google-adk
dotenv
llama-index